add changes for npm start
