+++ package.json
@@ -1,29 +1,46 @@
 {
   "name": "to-do-list",
   "version": "1.0.0",
-  "description": "This project is about simple to do list app using webpack.",
-  "main": "index.js",
+  "description": "A simple add and remove task.",
+  "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "webpack-dev-server --open",
     "build": "webpack"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/supreeti/To-Do-List.git"
+    "url": "git+https://github.com/supreeti/to-do-list.git"
   },
   "keywords": [],
-  "author": "",
-  "license": "ISC",
+  "author": "supreeti",
+  "license": "MIT",
   "bugs": {
-    "url": "https://github.com/supreeti/To-Do-List/issues"
+    "url": "https://github.com/supreeti/to-do-list/issues"
   },
-  "homepage": "https://github.com/supreeti/To-Do-List#readme",
+  "homepage": "https://github.com/supreeti/to-do-list/#readme",
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
+    "css-loader": "^6.7.3",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-html": "^7.1.0",
+    "eslint-plugin-import": "^2.27.5",
+    "hint": "^7.1.3",
     "html-webpack-plugin": "^5.5.0",
-    "webpack": "^5.76.0",
-    "webpack-cli": "^5.0.1"
+    "style-loader": "^3.3.1",
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.21.0",
+    "svg-inline-loader": "^0.8.2",
+    "webpack": "^5.76.1",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
+    "eslint-import-resolver-webpack": "^0.13.2",
     "lodash": "^4.17.21"
-  }
+  },
+  "main": "webpack.config.js"
 }
