--- webpack-demo/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "webpack-demo",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "watch": "webpack --watch",
-    "start": "webpack serve --open",
-    "build": "webpack"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "css-loader": "^6.7.3",
-    "html-webpack-plugin": "^5.5.0",
-    "style-loader": "^3.3.1",
-    "webpack": "^5.76.1",
-    "webpack-cli": "^5.0.1"
-  },
-  "dependencies": {
-    "lodash": "^4.17.21"
-  },
-  "main": "webpack.config.js",
-  "description": ""
-}
