--- webpack-demo/src/index.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>To Do List</title>
-    <link
-      rel="stylesheet"
-      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
-    />
-  </head>
-  <body>
-    <main class="main">
-      <header class="header">
-        <h1 id="title">Today's To Do</h1>
-        <svg xmlns="http://www.w3.org/2000/svg" height="24" width="24">
-          <path
-            d="M4.85 19.15v-.7h3.2L6.8 17.2q-1.125-1.125-1.637-2.438Q4.65 13.45 4.65 12.1q0-2.3 1.288-4.188Q7.225 6.025 9.35 5.15v.75q-1.825.8-2.912 2.487Q5.35 10.075 5.35 12.1q0 1.25.475 2.425Q6.3 15.7 7.3 16.7l1.25 1.25v-3.2h.7v4.4Zm9.8-.3v-.75q1.825-.8 2.912-2.487 1.088-1.688 1.088-3.713 0-1.25-.475-2.425Q17.7 8.3 16.7 7.3l-1.25-1.25v3.2h-.7v-4.4h4.4v.7h-3.2L17.2 6.8q1.125 1.125 1.637 2.438.513 1.312.513 2.662 0 2.3-1.288 4.187-1.287 1.888-3.412 2.763Z"
-          />
-        </svg>
-      </header>
-      <section class="todo_form">
-        <form id="form">
-          <div class="todo_wrapper">
-            <input type="text" placeholder="Add to your list..." id="task" />
-            <button class="btn" type="submit">
-              <svg xmlns="http://www.w3.org/2000/svg" height="20" width="20">
-                <path
-                  d="M8 15.25 4.25 11.5l3.729-3.729.313.291-3.23 3.23h8.98v-7.75h.416v8.166H5.062l3.23 3.25Z"
-                />
-              </svg>
-            </button>
-          </div>
-        </form>
-      </section>
-      <section>
-        <ul id="lists"></ul>
-      </section>
-      <section class="clear" id="allCompleted">
-        <p class="clearText">Clear all completed</p>
-      </section>
-    </main>
-  </body>
-</html>
