--- webpack-demo/src/style.css
@@ -1,110 +1,0 @@
-body {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-.main {
-  width: 40%;
-  margin: 100px auto;
-  border: 2px solid;
-}
-
-#title {
-  font-size: 20px;
-  margin: 0;
-  color: aqua;
-}
-
-input {
-  width: 100%;
-  border: none;
-  height: 50px;
-  font-style: italic;
-}
-
-input:focus {
-  outline: none;
-}
-
-input[type="checkbox"] {
-  margin-right: 10px;
-}
-
-input[type="checkbox"]:checked + input {
-  text-decoration: line-through;
-  color: #cacacb;
-}
-
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px;
-}
-
-.todo_form {
-  border-bottom: 1px solid #f4f5f4;
-  padding: 15px;
-}
-
-.todo_wrapper {
-  display: flex;
-  align-items: center;
-}
-
-.btn {
-  border: none;
-  background: none;
-  outline: none;
-  width: 25px;
-}
-
-.clear {
-  text-align: center;
-  background-color: #f6f6f6;
-  cursor: pointer;
-}
-
-.clearText {
-  margin: 0;
-  padding: 15px;
-  color: black;
-  background-color: antiquewhite;
-}
-
-.desciption {
-  width: 100%;
-}
-
-#lists {
-  padding: 0;
-  margin: 0;
-  list-style: none;
-}
-
-.listTask {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  border-bottom: 1px solid #f4f5f4;
-  padding: 10px;
-}
-
-.listTaskWrapper {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-#checkbox {
-  width: 15px;
-  border-color: #c0c0c3;
-}
-
-.todoBtn {
-  border: none;
-  background-color: white;
-  outline: none;
-  cursor: pointer;
-}
